Assignment 2
Task 1:

Create a basic page in html that displays an image.
When the user places a cursor over the image, replace the image with a different image.
When the user removes the cursor from the image, to return it to its original state.
Task 2:


Create a basic page in html that displays 3 unique images.
Create a separate rollover for each of these images, i.e.when mouse is over on any image it display a new, unique image and return it to the original image when mouse cursor is removed from the image.
Task 3:
Add a fourth image to your page.
The fourth image when moused over will not change. Instead, it will change the other three images on the page (these images have to be unique).
Then, onMouseOut of the fourth image, return the other 3 images to their original images.